Test for push
Finished all the tests for push. Again, based on the manual.

I cheated a little bit with how the flags is being stored,
though. That doesn't follow the manual and uses a more
convenient approach, so that part doesn't follow the manual.

Same for the pop. Forgot to add it to the message.