Test for LDAX
All of them based on examples on the manual.

I changed the urls, though, since the manual mentions memory
address for which I didn't build infrastructure yet:

I'm only considering the first 8kb of memory, i.e., the ones
where the software is installed.

In really, cpu should be able to access more. For now, I'm not
worrying about that to simplify stuff.