Add SUI instruction and tests
Created an implementation of the SUI instruction.

Created tests for the implementation of that instruction.

This gloriously found two errors:

1) The subtraction of u8 was made by complementing a u16 number
to handle the carry case. However, that would convert the eight
most significant bytes into ones, when they should remain as
zeros. I fixed it by complementing the u8 representation of
the number and then converting back to u16.

2) The logic for the parity bit was utterly wrong. I naively
thought it was meant to flag whether a result was even or not,
while instead it's to flag whether the number of set bits in a
result is even or not. I fixed it by adding a bit_count
function and checking the oddity of that number instead. The
function is probably not the fastest it could be (check [1]),
but it serves it's purpose for now.

[1] https://stackoverflow.com/questions/109023/how-to-count-the-number-of-set-bits-in-a-32-bit-integer