Added XRI instruction and tests
Added XRI instruction.

Added tests based on examples on the manual.

Changed the order of some arguments to make the name of the
variables match the name of the arguments in the function.

In practice doesn't matter because the operation (XOR) is
commutative, but OCD.

Also added ANI to the instruction match case. Forgot in
previous commit.