Finish(-ish) test for math
Finished all the tests for the math operations based on the
examples in the manual.

There's one notable exception: DAA. The manual doesn't provide
a good example to test, so I'll have to hand write one based on
what I think it does, which will of course pass in a
meaningless way.

A quick search in internet suggests that this isn't actually
used in this chip. I also found [1] which seems like a good
enough implementation in which I can inspire myself if I ever
need to support this instruction.

[1] https://github.com/begoon/i8080-core/blob/32d345da61796e9e740b9b6d303f5f4f849e7720/i8080.c#L521