Added support for HLT
Added support for the HLT instruction.

This basically meant adding a new field to the CPU: Status.

That field can right now be either Running or Stopped.

If the cpu is stopped, it doesn't run any instruction (no tests
for that yet) till an interruption is fired.

Otherwise, it works as normal.

HLT sets the status to Stopped.