Added support for IN/OUT
Added support for IN/OUT based on the manual.

This can be improved in a lot of ways, but it's a good start.

Some pain points for the future:

1) If you add a device to the cpu, you can't have another
reference to it as long as you have the cpu in scope. This may
become a pain in the ass with the borrow checker when
implementing the emulator of the full console. No idea yet on
how to solve it. Let's leave that problem for the future.

2) This implementation assumes that you'll plug devices
starting in the first possible port. That's not true in the
space invaders console, where the first output port is id 2.
This could be solved by changing the types of the vectors to
Vec<Option<Device>> instead of Vec<Device> and then adding a
bunch of Nones at the beginning. Not the cleanest, but good
enough without going crazy.