Make shell plug some devices
Make the shell of the console plug some devices to the cpu.

I didn't decide yet if it's good that the cpu is public in the
console object.

Also, I made the test for the OUT instruction kinda pointless,
as it doesn't guarantee that an assert is executed. Not sure
how to solve it. Yet! I feel that this is a place for closures,
though.