Implement timing
Do a basic implementation of the timing issue with the console.

Here we have a timer object that checks when it's time to fire
an interruption.

It also keeps track of the time elapsed. The console uses that
to calculate how many cpu cycles passed in the console and
executes them.

Also moved a bunch of stuff into the library and outside of the
main module, as Uncle Bob likes.