Make memory shareable
This commit includes the basic change to make the memory of
the console shareable.

Why did I choose Vec<Cell<u8>> instead of Cell<Vec<u8>>?

Excellent question. It's a trade off.

There's no simple way to edit a Cell<Vec<T>> without recreating
the whole vector in memory, modify the value and then replace
the cell with the new vector.

Because both seem equally wasteful memory wise, this option
seems to be the easiest one to implement and deal with.