Basic screen implementation.
Couple things:

1) Rename the old screen to printer. More appropriate.
2) Create a basic screen trait.
3) Create an implementation of that trait for the terminal.
4) Write down the logic of reading information from the memory
and map to screen.

Questionable decisions made:

1) I decided that the coordinate 0,0 in the layout will map to
the left, bottom corner of the screen. That makes math easier
now. I'm conscious that it may make math harder when printing
in the terminal, but it's 8am on a Saturday and I really don't
want to do math now.

It's a problem for future Agustin.