Split Keypad into two structures
One that acts as an interface between the player and the
console, and another one that is an interface between the
console and the CPU.

Questionable choices: Is the GameButton enum really needed?

One could argue it isn't. I like that abstraction, though, it
reduces my need to question whether enter was fire or start.