Fixed how frame buffer conversion
Found the error on the view!

The problem: I was doing the rotation completely wrong. Like
very very very wrong.

Fixed the algorithm. Also, to make it simpler to reason
about, I modified the interface to receive only the frame
buffer and not the whole memory. That way I don't have to go
around playing with address and I can assume all starts at zero
and goes till FRAME_BUFFER_SIZE.

The image is flipped. But that'll be solved in another commit.