Draw node function
Basically we have a node, which is composed by two integers:

x (the position in the x axis) with values between 0 and 223.
y (the position in the y axis) with values between 0 and 255.

And we need to modify the bit in the frame buffer that
represents that position to set it to one.

There're three steps:

1) We perform x * 0x20 + 2400 to get the address of the column
that stores the bit.
2) We divide the previous result by 8 to get the byte that
stores the bit.
3) We perform the rest of the result at 1) divided by 8 to get
the bit in the byte.

Then we set that bit to one (and only that one!) and store it
back.

Performing the rest of a division by a power of two is trivial:
Just do an & with n-1.

Performing the division by a power of two named n is possible
by using the external shift hardware:

We set the low byte to the value, the high byte to zero and the
shift to 8-x, where 2^x = n.

Performing the multiplication is trickier, because it happens
in two bytes and not only in one. But I suspect is possible by
using the external shift hardware too.

What you see there is my impulsive attempt to implement that
but I haven't actually cross checked that. Will explain that
once I actually understand it.