Fix multiplication by 0x20
How does this work?

The problem: We have a byte, we want to multiply it by 32 and
store the result in two bytes, one for the lower bits and
another one for the higher bits.

How do we do it?

First, note that 2^5 = 32. Knowing this, we can use the
external shift hardware to do the trick:

1) Set the shift offset of the hardware to 5.
2) Set the input for the shift to be the number we have to
multiply in the high order bits and 0 in the low order bits.
3) The output of that shift is the lower order bits.
4) Now set the input to be the number we have to multiply in
the lower order bits and 0 in the higher order bits.
5) The output of the shift is the higher order bits of the
result.