Implement the logic of RESUME_INIT
The idea: If there's an interruption while the game has been
initializing, we want to come back to work as soon as possible
where we were before.

To do so, we check the steps register. 0x04 means the game is
running. If it's not that, it means initialization is still
happening. Let's come back to work then!

Otherwise, let's update values in screen.