Save state on interruption
Save the state of the cpu on interruptions, to make sure that
we return exactly where we were.

This wasn't important before the implementation of the timer,
since the only place that had sensible status was the drawing
routines (and they can't be interrupted) and the initialization
steps (and on interruption we return to them right away).

The timer, though, has some status that we need to recover.