Fix the problem when changing direction
We don't need a new vertex!

The problem was the following. Suppose we have a worm that's
going up:

_ _ _
_ X _
_ X _
_ _ _

And then you change directions to go right. Because it'd update
first the head, after CALL DRAW_NEXT STEP, we would have:

_ _ _
_ _ X
_ X _
_ _ _

Now, UPDATE_TAIL will check first if the column is different
and then short circuit. But that would cause:

_ _ _
_ _ X
_ _ X
_ _ _

Which is not what we want!

At the beginning I thought we should add, somehow, a new vertex
and end with something like this:

_ _ _
_ X X
_ X _
_ _ _

But that would require a lot of weird logic. Instead, if we
update the tail before drawing the new node, this would happen:

At the beginning:

_ _ _
_ X _
_ X _
_ _ _

After UPDATE_TAIL:

_ _ _
_ X _
_ _ _
_ _ _

After DRAW_NEW_NODE:

_ _ _
_ X X
_ _ _
_ _ _

Exactly what we want!