Sketch of the cpu abstraction
Created a first draft of it.

A couple concerns: I had to make some methods more visible than
they originally were. _I think_ that's ok. Also, the
Instruction is not object safe. Not sure if that'll byte me in
the ass later.

Maybe I can make it so by requiring an implementation of From?