Implement methods of the cpu
Start implementing the basic methods of the cpu.

A couple side changes:

MOS 6502 doesn't have ports (interesting, eh!). So it doesn't
make sense for the Cpu interface to enforce that. I created a
new interface for cpus that have ports and made the Intel 8080
cpu implement it.

Also, interesting bit about rust:

Suppose you have a constant of one type and variable, immutable
of another. And they're called c and v respectively.

This code:

c as v_type == v

Doesn't output the same as:

c = v as c_type

In fact, the first option is slightly more optimized. I can
kinda understand it, but it's rather unfortunate. Anyway, I
changed the "is_done" function on the Cpu implementation for
the Intel 8080 to use the first alternative.

See this for an example of what I mean:

https://play.rust-lang.org/?gist=7e458d44fcc2121c101440e298e15430&version=stable&mode=debug&edition=2015