First execute, then get cycles
Change the order in which this happens. Rationale:

The MOS 6502 depends a lot on the status of the cpu to know how
many cycles an instruction performed, including stuff that
isn't only the flags. That logic would get duplicated with the
current setup. I'd need some kind of new status to flag
conditions being met (or not). That status would be updated in
the execution of the instruction and consulted when asking for
the cycles executed.

The previous order didn't support that and required the status
to be updated while asking for cycles, which is stupid.