Remove panic from internal function
get_value_from_addressing_mode had a panic! that could easily
be replaced by a `CpuError`. So a bit of refactor to make it
work that way.

Also, after doing that it became possible to use
get_address_from_addressing_mode instead of having the logic to
calculate the memory addressed duplicated in two parts of the
code. So I dedupped that.

And lastly, in the commit I made past night from bed from the
github UI to the space_invaders/main.rs file there was,
unsurprisingly, a mistake. So I also fixed that.