Implement SAX
And fix how I was checking addressing modes.

The fact that an undocumented instruction is the same than the
concatenation of other two instructions, doesn't mean that
those two instructions have compatible possible addressing
modes.

To address this fact, I separated the concept of executing an
instruction and checking whether the addressing mode was valid
or not. I'm not extremely proud about the naming, but will
remain like that till I think of something else.