Actually make tests run
A couple errors here:

- The index in argv we were accessing to get the file name was
wrong.
- The conversion of AVAILABLE_MEMORY to u16 was causing the
result to be zero (makes sense! AVAILABLE_MEMORY > 2^16), so
the CPU was never starting. Instead of converting the constant
to the type of the register, I decided to cast the register to
the type of the constant, which is a safer conversion since it
goes to an integer with more bits, and not to one with less.