Fix error on how to branch back.
I was assuming (VERY WRONGLY) that the byte passed as an
argument to the Relative addressing mode was in the same format
as the i8 number in Rust.

Big mistake. i8 is a signed binary number and the number passed
to Relative when wanting to branch back is two complement. Fix
that and add a test, please.