Fix the break flag
Apparently, it doesn't work as other flags. From
./docs/6502.txt:

```
B Break flag
This flag is used to distinguish software (BRK)
interrupts from hardware interrupts (IRQ or NMI). The B
flag is always set except when the P register is being
pushed on stack when jumping to an interrupt routine to
process only a hardware interrupt.

The official NMOS 65xx documentation claims that the BRK
instruction could only cause a jump to the IRQ vector
($FFFE). However, if an NMI interrupt occurs while
executing a BRK instruction, the processor will jump to
the NMI vector ($FFFA), and the P register will be
pushed on the stack with the B flag set.
```

A consequence of this is that the break flag isn't affected by
PLP or PHP. And also that I have to somehow implement IRQ and
NMI, which weren't documented in the page I used as a reference