Fix how JSR and RST work
So... My code isn't necessarily broken, it works, but it's not
complaint with the Klaus test suit.

What happens?

Well, apparently the MOS 6502 specifies that JSR should save in
the stack PC-1 and RST should restore *(STACK POINTER)+1.

Me, being the eager optimizer that I'm, decided that instead I
would just store PC itself and restore it itself, saving two
arithmetic operations and effectively achieving the same.

However, Klaus tests whether the exact implementation happens.

Which it's... fair. My optimization relied on the assumption
that people wouldn't be stupid and modify the value of PC in
the stack under the assumption that RST will later add one.

That's a big assumption.