Fix overflow for SBC
Apparently, when substracting -128 and zero, we should set the
overflow flag. Which seems rather wrong, according to the
definition of overflow I saw everywhere (i.e. when the results
doesn't fit in the range [-128, 127]).

So we need this specific logic for SBC. I tried to use said
logic in ADC too, but it didn't work and I have no patience to
understand why.

This chip is weird, man.