Handle better edge cases in SBC
Soooo, this code has a problem:

If you have carry=false, A=1, M=0, then the result is 100.
Which will cause the carry flag to be set. Which is not what we
want.

I have the feeling that there's a more elegant solution to
this. But since I have no idea so far, let's leave it here.