Fix SBC computation
Correction: It's not that carry should be set when carry=true,
M=0 and A=0, it's that carry should be set when there's _not_
an overflow. What happens: I was hoping that the two_complement
subtraction would solve the problem reasoning about that by
inverting the issue. And it does, but not for that specific use
case.

Frustrated, I found a beautiful function in the Rust API:
overflowing_sub. This function handles most of this weird logic
in a way that's sensible to my use case.

Very nice edge case spotted by the tests, to be fair. Very
nice. And very nice functions found in Rust!