Fix how decimal addition works
And obviously there was an error in the doc :).

Basically, that condition is to check whether the sum of the
first digits should "add one" to the second digit (i.e. the sum
is bigger than 9). We were doing that logic as if this were hex
addition and not decimal.