Create a memory trait
Instead of using [0; S] as the memory, create a trait to
abstract the operations to do with it and use that instead.

Implement the trait for [0;S] so the existing tooling works as
usual.

Why is this needed?

The memory ram used by the NES is special, in the sense that
there're a lot of logical mirrors. In order to implement that,
I'd have to either couple those concepts into the current CPU,
which is bad if I want to use it in other emulators, or build
some abstraction to handle CPU. None of those seem really
appealing. This will allow me to implement my own RAM in the
NES that handles that logic internally. Hopefully, that'll be
enough.