First draft of NES memory
Based on https://nesdev.com/NESDoc.pdf, pages 10 and 11.

I'm not really sure about the implementation of slice. It
assumes that the stack pointer will always be somewhere in the
rom... Which _I hope_ is right, because it keeps it simple. But
opens some questions:

- What happens if that's true but someone passes an invalid
index to the function? Right now it'll just overflow. Should we
refactor slice to return Result? That might be needed for the
set function too: Right now it has no behaviour if the user
tries to set the rom. But wouldn't an error be a better
behavior?

- What happens if that's _not_ true and there're games in which
the stack pointer can go outside the rom? I can see that
happening... In that case we would have to refactor slice to
smartly grab the slice from where it should. It can become
pretty pretty quickly though. So let's hope (fingers crossed!)
I never have to that.

Which also brings the question... What should the initial value
of the stack pointer be? 0x8000? Let's keep reading the
documentation and see.