Create register 2000
Basically:

The PPU has a register. That register has a connector. That
connector has also a reference to the register. The RAM has
access to the connector. We _could_ probably simplify it by
making the connector and the register one same thing, but that
would require the RAM to have a
Option<Rc<RefCell<InputOutputDevice>>> instead of the current
Option<Box<InputOutputdevice>> which seems overlycomplicated.