Remove unnecessary boxing
We just pushed mutability down the abstraction tree, so we can
remove it from that vector.

Ideally, we'd use an array with fixed type, but because we have
that ugly device type in IORegister, we can't make it derive
Copy and therefore be able to use the array.

Well, life isn't perfect.