Remove the slice function
So what happens.

In order to make the MOS 6502 and the PPU share the ram, I need
to implement Memory for Rc<RefCell<T: Memory>>.

The implementation of Memory::slice for that type isn't trivial
as `self.borrow.slice(from,to)` wouldn't pass the borrow
checker (basically, the lifetime of the reference returned by
slice can't be proved to be wider than the lifetime of self).

So we need to get rid of slice. To do that, we need to change
`get_next_instruction_byte` to return a Vec<u8> instead of a
slice. That's ok, because we were converting into a vector
anyway.

As a side effect of the changes in the NES structure (i.e. add
a cpu and make it share ram with the ppu), we also need to
change the type of the memory in the cpu from a mutable
reference to a trait object. That's not ideal, but it's also
ok.