Fix some things in the Snake game
While re-reading, I noticed some errors:

1. I refactored how the status register (B) works during
initialization to use less bytes. In this case, instead of
keeping duplicating the value, I'd just move the finish flag
(0x08) once it's finished.

2. Changed the order in which initialization happens: Instead
of init the snake and then clearing the screen, I first clear
the screen and then init the screen. The reason for this change
is to fix another error: We were printing only one of the two
nodes on initialization! This way I don't have to duplicate
logic between initialization of snake and initialization of
screen to print both nodes. I just make both happen at the same
time.

3. Fixed the wording of a method that was impossible to
understand.