NES has decimal mode disabled
Allow for a version of the MOS 6502 with decimal mode disabled
and make the NES use that CPU instead of the one with the mode.

We could argue that in theory we don't need it: _Why_ would
developers use a flag for a mode that's disabled. But since
both CLD and SED work as normal and there are games that do
nasty things such as stack manipulation+RTS instead of RTI,
it's better to not assume developers are sane.

Also fix a typo in the doc.