Fix snake data structure
The first node in the linked list should be the head.

That means that if the first position is:

- - -
- x -
- x -
- - -

On the center of the screen, the node on the top should be the
first node. That was not the case.

Also, we want to draw first and then update the tail.

That's because on the initial setup, if we don't do that, we
would advance the tail, it would overlap with the head and the
system will wrongly remove the head node. However, if we first
update the head, that wouldn't happen because they won't be
the same.