Move how often the snake moves
Initially, we were doing it the dumb way:

- Moving it on every hardware interruption.
- Updating the direction every half second.

That meant that the speed of the snake will be 60 times faster
than the speed of the input reads. Which is ridiculous. Instead
let's make them the same by doing both every half second.

To do so, I also need to stop using B to hold the init status.

Instead, Let's use memory.