Parse more instruction with the new schema
Also noticed something: The jump instructions were being parsed assuming
that the assembly code for branch instructions will be `INT NUM NUM`
where `NUM` are bytes. However, it'd be `INT NUM` where `NUM` is a
double word. So I made a couple adjustments regarding that.

Also, because of how the lexing work (if it fits, it's a byte), I'll
probably have to change the way parsing works for `DW` and `ORG` so it
accepts words as the next token and converts them to double words.

That's because in the statement `ORG 0`, the lexer will convert
represent `0` in the parsing tree as a Byte. Then the parser will be
expecting a double word and will consider the statement malformed, even
though it isn't.