Make the room a vector
This is the first step in a serie of changes to allow to use labels that
are going to be declared later.

How it will work:

There will be a first pass that will materialize all it can, but the
labels into a first stage room.

Then there will be a second pass that it will grab all the labels in the
first stage room, convert them to bytes and store them in the final room.