Use a pointer to props instead of props in object
Instead of having a reference to the properties array as part of the Object Value,
lets store the place in memory in which the reference to the properties array is.

This prevents the case of having to resize the object when adding a property from
creating a completely different object and having to update all references.